Created ft_itoa.c it works and respects the Norm. modified strlcpy

+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: ameechan <ameechan@student.42.ch>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/14 15:14:27 by ameechan          #+#    #+#             */
-/*   Updated: 2023/10/20 14:20:25 by ameechan         ###   ########.fr       */
+/*   Updated: 2023/11/02 16:46:35 by ameechan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t dest_size)
 {
-	unsigned long	i;
-	int				src_len;
+	size_t	i;
+	int		src_len;
 
 	i = 0;
 	src_len = ft_strlen(src);
